feat/LIVE-2799-wallet-page-transactions-list section container added to asset page
