+++ apps/ledger-live-mobile/src/screens/Portfolio/index.tsx
@@ -11,10 +11,9 @@ import { useTranslation } from "react-i18next";
 import { useFocusEffect } from "@react-navigation/native";
 import { isAccountEmpty } from "@ledgerhq/live-common/account/index";
 
-import { Box, Flex, Link as TextLink, Text } from "@ledgerhq/native-ui";
+import { Box, Flex, Link as TextLink } from "@ledgerhq/native-ui";
 
 import styled, { useTheme } from "styled-components/native";
-import proxyStyled from "@ledgerhq/native-ui/components/styled";
 import { PlusMedium } from "@ledgerhq/native-ui/assets/icons";
 import { useRefreshAccountsOrdering } from "../../actions/general";
 import { accountsSelector } from "../../reducers/accounts";
@@ -44,6 +43,7 @@ import CheckTermOfUseUpdate from "../../components/CheckTermOfUseUpdate";
 import TabBarSafeAreaView, {
   TAB_BAR_SAFE_HEIGHT,
 } from "../../components/TabBar/TabBarSafeAreaView";
+import SectionTitle from "../WalletCentricSections/SectionTitle";
 import OperationsHistorySection from "../WalletCentricSections/OperationsHistory";
 
 export { default as PortfolioTabIcon } from "./TabIcon";
@@ -60,71 +60,12 @@ type Props = {
   navigation: any;
 };
 
-const StyledTouchableOpacity = proxyStyled.TouchableOpacity.attrs({
-  justifyContent: "center",
-  alignItems: "flex-end",
-  px: 7,
-  mx: -7,
-  py: 5,
-  my: -5,
-})``;
-
 const SectionContainer = styled(Flex).attrs((p: { px?: string | number }) => ({
   mt: 9,
   px: p.px ?? 6,
 }))``;
 
 export const Gradient = styled(BackgroundGradient)``;
-
-const SectionTitle = ({
-  title,
-  onSeeAllPress,
-  navigatorName,
-  screenName,
-  params,
-  navigation,
-  seeMoreText,
-  containerProps,
-}: {
-  title: React.ReactElement;
-  onSeeAllPress?: () => void;
-  navigatorName?: string;
-  screenName?: string;
-  params?: any;
-  navigation?: any;
-  seeMoreText?: React.ReactElement;
-  containerProps?: FlexBoxProps;
-}) => {
-  const { t } = useTranslation();
-  const onLinkPress = useCallback(() => {
-    if (onSeeAllPress) {
-      onSeeAllPress();
-    }
-    if (navigation && navigatorName) {
-      navigation.navigate(navigatorName, { screen: screenName, params });
-    }
-  }, [onSeeAllPress, navigation, navigatorName, screenName, params]);
-
-  return (
-    <Flex
-      flexDirection={"row"}
-      justifyContent={"space-between"}
-      alignItems={"center"}
-      {...containerProps}
-    >
-      <Text variant={"h3"} textTransform={"uppercase"} mt={2}>
-        {title}
-      </Text>
-      {onSeeAllPress || navigatorName ? (
-        <StyledTouchableOpacity onPress={onLinkPress}>
-          <TextLink onPress={onLinkPress} type={"color"}>
-            {seeMoreText || t("common.seeAll")}
-          </TextLink>
-        </StyledTouchableOpacity>
-      ) : null}
-    </Flex>
-  );
-};
 
 const maxAssetsToDisplay = 5;
 
@@ -254,7 +195,10 @@ function PortfolioScreen({ navigation }: Props) {
         : []),
       ...(showAssets
         ? [
-            <OperationsHistorySection accounts={assetsToDisplay} />,
+            <SectionContainer px={6} mb={8}>
+              <SectionTitle title={t("analytics.operations.title")} />
+              <OperationsHistorySection accounts={assetsToDisplay} />
+            </SectionContainer>,
           ]
         : []),
     ],
@@ -296,7 +240,11 @@ function PortfolioScreen({ navigation }: Props) {
         />
         <AnimatedFlatListWithRefreshControl
           data={data}
-          style={{ flex: 1, position: "relative", paddingTop: 48, marginBottom: 36 }}
+          style={{
+            flex: 1,
+            position: "relative",
+            paddingTop: 48,
+          }}
           contentContainerStyle={{ paddingBottom: TAB_BAR_SAFE_HEIGHT }}
           renderItem={({ item }: { item: React.ReactNode }) => item}
           keyExtractor={(_: any, index: number) => String(index)}
