+++ apps/ledger-live-mobile/src/screens/WalletCentricAsset/index.tsx
@@ -2,6 +2,9 @@ import React, { useMemo } from "react";
 import { FlatList } from "react-native";
 import Animated from "react-native-reanimated";
 import { useSelector } from "react-redux";
+import { useTranslation } from "react-i18next";
+import styled from "styled-components/native";
+import { Flex } from "@ledgerhq/native-ui";
 import { getAccountCurrency } from "@ledgerhq/live-common/src/account";
 import accountSyncRefreshControl from "../../components/accountSyncRefreshControl";
 import { withDiscreetMode } from "../../context/DiscreetModeContext";
@@ -9,6 +12,7 @@ import TabBarSafeAreaView, {
   TAB_BAR_SAFE_HEIGHT,
 } from "../../components/TabBar/TabBarSafeAreaView";
 import { accountsSelector } from "../../reducers/accounts";
+import SectionTitle from "../WalletCentricSections/SectionTitle";
 import OperationsHistorySection from "../WalletCentricSections/OperationsHistory";
 
 type RouteParams = {
@@ -20,11 +24,18 @@ type Props = {
   route: { params: RouteParams };
 };
 
+
+const SectionContainer = styled(Flex).attrs((p: { px?: string | number }) => ({
+  mt: 9,
+  px: p.px ?? 6,
+}))``;
+
 const AnimatedFlatListWithRefreshControl = Animated.createAnimatedComponent(
   accountSyncRefreshControl(FlatList),
 );
 
 const AssetScreen = ({ route }: Props) => {
+  const { t } = useTranslation();
   const accounts = useSelector(accountsSelector);
   const { currencyId } = route?.params;
   const cryptoAccounts = useMemo(
@@ -33,7 +44,12 @@ const AssetScreen = ({ route }: Props) => {
   );
 
   const data = useMemo(
-    () => [<OperationsHistorySection accounts={cryptoAccounts} />],
+    () => [
+      <SectionContainer px={6} mb={8}>
+        <SectionTitle title={t("analytics.operations.title")} />
+        <OperationsHistorySection accounts={cryptoAccounts} />
+      </SectionContainer>,
+    ],
     [cryptoAccounts],
   );
